% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanjistat.data-package.R
\docType{data}
\name{dct}
\alias{dct}
\title{The component-transport distance matrix}
\format{
A 2136 x 2136 symmetric numeric matrix of distance values
between 0 and 0.25 with row and column names specifying
the corresponding jouyou kanji in the usual order used in the kanjistat
package, e.g. for \link[kanjistat:kanjidata]{kanjistat::kbase}.
}
\source{
The matrix has been generated by computing the individual entries
in kanjistat v0.13.1 as

\if{html}{\out{<div class="sourceCode">}}\preformatted{   kanjidist(k1, k2, compo_seg_depth1 = 4, compo_seg_depth2 = 4,               
             p = 1, C = 0.2, approx = "pcweighted", type = "rtt") 
}\if{html}{\out{</div>}}

where k1 and k2 are kanjivec objects of the two kanji considered.
}
\usage{
dct
}
\description{
A matrix of pairwise distances between jouyou kanji based on visual similarity
computed with the \code{\link[kanjistat:kanjidist]{kanjistat::kanjidist()}} function.
}
\section{Warnings}{


While the distance matrix seems to be useful for obtaining sets of similar
looking kanji (see the Examples below), there is a number of caveats:
\itemize{
\item Neither the distances nor the neighborhoods of similar looking kanji
are very useful for several of the kanji composed of very few
strokes. This is mainly because the \code{\link[kanjistat:kanjidist]{kanjistat::kanjidist()}} code
does currently not consider individual strokes as components.
\item In some cases the neighborhoods are inadequate due to
a suboptimal decomposition of the kanji (suboptimal
for the task at hand, that is). This is sometimes already based in the
structure of the kanjiVG file, but more often due to a lack of refinement
in the processing of special cases by \code{\link[kanjistat:kanjivec]{kanjistat::kanjivec()}}.
\item The distances do not satisfy the triangle inequality, i.e. we find many
combinations of i,j and k such that \code{dct[i,j] <= dct[i,k] + dct[k,j]} does
not hold. This is by design as the underlying component-transport distance
is not a metric.
}

While the first two points are clearly undesirable and will be improved on
in the future, the last point helps in the construction of the component-transport
distance function and is backed by research in cognitive psychology indicating
that human similarity judgment are better described by non-metric structures
(e.g. a strong feeling of similarity is triggered when a certain threshold in
"objective" similarity is passed). See Remark 1 of the reference below.
}

\examples{
dct[700:708, 700:708]
sort(dct[356,])[1:20]
}
\references{
Dominic Schuhmacher (2023).\cr
Distance maps between Japanese kanji characters based on hierarchical optimal transport.\cr
ArXiv, \doi{10.48550/arXiv.2304.02493}
}
\keyword{datasets}
