% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanjistat.data-package.R
\docType{data}
\name{dembed}
\alias{dembed}
\title{The embeddings distance matrix}
\format{
A 2136 x 2136 symmetric numeric matrix of distance values
between 0 and 0.25 with row and column names specifying
the corresponding jouyou kanji in the usual order used in the kanjistat
package, e.g. for \link[kanjistat:kanjidata]{kanjistat::kbase}.
}
\source{
Generated by Lennart Finke using \href{https://platform.openai.com/docs/guides/embeddings}{OpenAI's text embeddings}.
Python code https://github.com/dschuhmacher/kanjistat.data/tree/main/data-raw
}
\usage{
dembed
}
\description{
A matrix of pairwise distances between jouyou kanji based on similar semantics
computed using OpenAI's text embeddings.
}
\details{
TO DO.
}
\examples{
dembed[100:108, 100:108]
sort(dembed[158,])[1:20]
}
\keyword{datasets}
