% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanjistat.data-package.R
\docType{data}
\name{kvec}
\alias{kvec}
\alias{kvecjoyo}
\alias{kvecjinmei}
\title{Data in \code{kanjivec} format}
\format{
Each list element is an object of the S3 class \code{kanjivec}, see \code{\link[kanjistat:kanjivec]{kanjistat::kanjivec()}}.
}
\source{
The lists have been generated by \code{\link[kanjistat:kanjivec]{kanjistat::kanjivec()}}
with parameter \code{flatten="intelligent"} under kanjistat version 0.13.0 (\code{kvecjoyo})
and 0.13.1 (\code{kvecjinmeiyou}) based on the KanjiVG database by Ulrich Apel
\url{https://kanjivg.tagaini.net/}.
}
\usage{
kvecjoyo

kvecjinmei
}
\description{
\code{kvecjoyo} is a list of all 2136 current jouyou kanji (as last specified by the Japanese Ministry
of Education in 2010).

\code{kvecjoyo} is a list of those 774 out of all 863 current jinmeiyou
(as last adapted by the Japanese Ministry of Justice in 2017).
}
\examples{
plot(kvecjoyo[[869]])
plot(kvecjoyo[[869]], type="dend")
kvecjoyo[[869]]
str(kvecjoyo[[869]])
kanjistat::get_strokes(kvecjoyo[[869]], 10)  # the tenth stroke of the kanji 

}
\keyword{datasets}
